le composable de extract la partie qui communique avec fire base charger les donnée GET
